+++ packages/ragbits-core/src/ragbits/core/prompt/exceptions.py
@@ -8,12 +8,28 @@ class PromptError(Exception):
         self.message = message
 
 
-class PromptWithImagesOfInvalidFormat(PromptError):
+class PromptWithAttachmentOfUnknownFormat(PromptError):
     """
-    Raised when there is an image attached to the prompt that is not in the correct format.
+    Raised when there is a file with an unknown format attached to the prompt.
     """
 
-    def __init__(
-        self, message: str = "Invalid format of image in prompt detected. Use one of supported OpenAI mime types"
-    ) -> None:
-        super().__init__(message)
+    def __init__(self) -> None:
+        super().__init__("Could not determine MIME type for the attachment file")
+
+
+class PromptWithAttachmentOfUnsupportedFormat(PromptError):
+    """
+    Raised when there is a file with an unsupported format attached to the prompt.
+    """
+
+    def __init__(self, mime_type: str) -> None:
+        super().__init__(f"Unsupported MIME type for the attachment file: {mime_type}")
+
+
+class PromptWithEmptyAttachment(PromptError):
+    """
+    Raised when there is an empty file attached to the prompt.
+    """
+
+    def __init__(self) -> None:
+        super().__init__("Attachment must have either bytes data or URL provided")